ish: Add panic support for x86
Adds the right register mapping for i386. While the logs don't include
the info, this seems to be happening because of where the crash is.
The error is:

Fatal error: 0
  eax      = 0xFF21E292
  ebx      = 0xFF21EC71
  ecx      = 0x00000053
  edx      = 0xFF21E48F
  esi      = 0xFF209A2B
  edi      = 0x00000001
  cs       = 0x00010008
  eip      = 0xFF209104

This maps to a libc function `__m_vfprintf` which would affect
logging. I was able to get around it by temporarily replacing the
log format with manual hex to string and printing raw strings.
We'll fix the crash in b:399264195.

BUG=b:398737082
TEST=Run on Trulo. This still doesn't print the log but the info is
     in the right place now.

Change-Id: I3e8ff0f7c2e6051083a954445afb6d03d65294a8