ppm: Query ports from `named-usbc-port` nodes
The PPM driver currently accepts a list of PDC node phandles in the
device tree in order to tell it which PDCs to interface to when
executing UCSI commands.

However, this is unnecessary and can lead to inconsistencies where port
indexes and counts could vary between the PPM/UCSI and `pdc_power_mgmt`.
Instead, let the PPM driver reference the `named-usbc-port` nodes and
get the PDC driver references from there. This makes `named-usbc-port`
nodes the sole source of truth.

BUG=b:372910804
TEST=./twister -T zephyr/test/pdc

Change-Id: I13afb4e04f2497cdb5a8536b9451cda593dd2e13