prochot-vcmp: Add support for asserting PROCHOT
For systems that use a single pin to monitor PROCHOT and assert PROCHOT
to the AP, add an optional child-binding to the prochot-vcmp node that
configures the GPIO and pinctrl settings needed to automatically switch
the PROCHOT between ADC and GPIO mode.

BUG=b:214509787
TEST=Enable AP throttle on brox.  Use the "thermalset" console command
to force a PROCHOT condition.  Observe that the CPU speed throttles to
400 MHz during PROCHOT and resumes normal speed once PROCHOT is cleared.

Change-Id: Ib33fc71181bef37834e398b2a846d4007055cbc0