emul_tcpci_partner: Implement FRS
Trivially implement Fast Role Swap to use the same message sequence and
timing as Power Role Swap. Provide a mechanism for the partner emulator
to signal FRS to the TCPC emulator, which will then trigger an alert
interrupt. Model failure to respond to FR_Swap as a faulty partner
variation and not the default behavior. Refactor FRS tests to use these
facilities. Support FRS tests in which the partner completes the FRS
sequence.

BUG=b:361800172
TEST=twister -s zephyr/test/drivers/drivers.usbc_frs

Change-Id: I1335140e14fec6b84dbd4dcf039a6f0d92f2fe2d