WIP: nissa: guard USB-C port count
Nissa boards reference a "sub-board type" field in the CBI FW_CONFIG
value to determine their USB-C port count, which can vary between 1 and
2 depending on the board. Some Nissa boards define only 1 USB-C port /
TCPC in their devicetree because that model only ever supports one port.
However, this can cause a crash if the CBI data is misprogrammed and
fools the EC into thinking there is a second USB-C port present.

As a temporary solution, lock the port count to 1 if
CONFIG_USB_PD_PORT_MAX_COUNT is 1. If there is hardware for more than
one USB-C port, then consult CBI.

BUG=b:400821012
TEST=zmake build nivviks pujjo
TEST=Tested on a Pujjo DUT
LOW_COVERAGE_REASON=Coverage to be updated in b/402282721

Change-Id: Ia46438af3e06757a7fa6454aecc68fe4c4637bab