board/hyperdebug: Switch JTAG clock phase
The JTAG phase is changed such that the clock is high when idle, that
is, leading edge will be falling and trailing edge will be rising.

The JTAG device being debugged will still latch TMS and TDI on the
rising edge of the clock, which is now the trailing edge.  This
sidesteps any issue around the steady clock level before a burst of JTAG
cycles.

BUG=b:310714758
TEST=./bazelisk.sh test //sw/host/provisioning/orchestrator/tests:e2e_emulation_tpm_hyper310_test  (in third_party/lowrisc/opentitan branch earlgrey_1.0.0)

Change-Id: Id618a0d148b710ce6b3d81ed150628d282d2886b