zmake: Logging cleanup, enable job name logging for multi-proj builds
zmake is able to to prepend job names like "[skyrim:rw]" in front of log
lines to indicate what project is currently being configured/built/etc.
Currently, this is only enabled when `--log-label` is set (which also
turns on level printing). Seeing the job names is especially useful in
CI logs when we need to determine which project(s) experienced errors.

Make these changes to the logging:
  * For compare-builds, build, and configure, turn on job name logging
    if more than one project is selected so log output can be
    distinguished.
  * Also print job names from the output packer stage
  * Make some other log messages consistent by moving the project names
    up to the front of the line. Use a space between the [job-name] and
    log message
  * Refactor the LOG_JOB_NAMES global variable into a class variable
  * Update unit tests

Why not just add `--log-label` to the zmake call in
`firmware_builder.py`? This also adds level labels to the logging
("ERROR: ", etc) on every line, which are frequently wrong. For example:

```
ERROR: [brox:ro] WARNING: Unknown or disabled ...
ERROR: [brox:ro] INFO: Project brox does not s...
ERROR: [brox:ro] INFO: No motionsense-sensor-a...
```

This adds way to much noise to the log and requires additional work to
clean up. So use just the job names for now.

BUG=None
TEST=zmake build <projects>
TEST=zephyr/zmake/run_unit_tests.sh

Change-Id: I30034cd3ea2c5539e0348f68d1f5caf71a6ecf6e