dirks: Implement board_is_power_good function
This patch implements the board_is_power_good function to report
the power failure when the system supply power not enough.
If the system powered by BJ adapter, it checks gpio_bj_adp_present.
If the system powered by USBC adapter, it checks if the actual VBUS
is more than 5% less of the requested VBUS.

BUG=b:399236160
TEST=verify on Dirks with 45W/65W USBC adapter and 65W BJ adapter
     1. removing AC power when system is on
     2. reconnect AC power and the AP boots without pressing power
        button.
     ./twister -T zephyr/test/nissa/ -v -i -s nissa.dirks PASS

Change-Id: I329a199e2b1a7e20de96914dffa2a159edea613a