nds32: Handle debug exception (DEX) mode
When `break` is invoked, nds32 enters debug exception mode. While in
this mode interrupts are disabled. This is not reflected in the usual
register used to check interrupt and context status. This can cause a
deadlock and watchdog hang since the system assumes it is safe to wait
for a mutex or an interrupt.

This change effectively treats DEX mode the same as executing in
interrupt context with interrupts disabled.

BUG=b:408066226
TEST=`crash assert` does not hang
TEST=`crash stack` does not hang
TEST=`crash watchdog` contains a PC

Change-Id: I8617a99e5aba82dd6cc10321aedc86a3a3539e54