zephyr/fatcat_it82002aw: Initialize EDP backlight enable high
Fixes an issue where the backlight for the EDP panel on the
fatcat_it82002aw board was off by default, making the screen
unreadable.

This commit modifies the `gpio_c_edp_bklt_en_r` GPIO definition
in the generated DTS to include the `GPIO_OUTPUT_INIT_HIGH` flag.
This ensures that the backlight enable pin is initialized to a high
state, turning the backlight on by default.

TEST=Able to boot google/fatcatite w/ active display.

Change-Id: I478ff3770dca9b4d352f5ecd1155a2eba9f7db33
Signed-off-by: Subrata Banik <subratabanik@google.com>