cortex-m: Clean up exception stack and mode handling
The watchdog_trace handler was only checking for 0xfffffff1 exceptions
when determining which stack the exception frame is in. This resulted
in a bad PC being populated in the panicinfo, often being the stack
canary deadd00d. The check should be for 0xfffffff1 or 0xfffffff9.

Since is_frame_in_handler_stack in cortex-m/panic.c already performs
this check, it was  moved to cpu.h so both watchdog.c and panic.c
can utilize it.

is_exception_from_handler_mode was also added for checking which mode
the exception occurred in. Cleaned up a few places where exception stack
or mode was being checked.

BUG=b:405199532,b:378342175
TEST=PC correct on watchdog warning

Change-Id: I01c7c74b964811089bfb51c0c174ef40a3180ba2