cortex-m: Handle 0xfffffff9 exceptions in watchdog
The watchdog_trace handler was only checking for 0xfffffff1 exceptions
when determining which stack the exception frame is in. This resulted
in a bad PC being populated in the panicinfo, often being the stack
canary deadd00d. The exception frame is on the main stack when the
exception return/lr is 0xfffffff1 or 0xfffffff9.

Since there was already a function doing this check in cortex-m/panic.c,
it was moved to cpu.h so both watchdog.c and panic.c can utilize it.

BUG=b:405199532,b:378342175
TEST=PC correct on watchdog warning

Change-Id: I01c7c74b964811089bfb51c0c174ef40a3180ba2