common/tablet_mode: handle keyboard suspend and resume
In scenarios where DSP communication is enabled, keyboard
scan should be suspended and resumed when the system's
power state changes.
changes allow this functionality when the DSP communication
inetrface is enabled and there is no tablet mode interrupt
handler defined on EC side.

BUG=b:409240881
TEST=zmake build trulo-ti
In tablet mode suspend the DUT.
While DUT is in suspend mode, press any key on the onboard
keyboard. The DUT does not waking up.

Change-Id: I120a7895703635e91e39b93f2ae228bf906630e4