pdc: rts5453: Add console-based PDC FW update mechanism
To allow engineers to flash the RTK PDC on boards without needing to
prepare AP FW images for development purposes, create a mechanism to
stream FW updates over the EC console from a host. See subsequent CL
for a Python script that is able to interact with this new CLI interface
(pdc fwup_rtk) and transfer a PDC FW image.

This requires that USBC_PDC_RTS54XX_CONSOLE_FW_UPDATER be enabled.

BUG=None
TEST=Successfully reflashed PDC

Change-Id: I62d917023ad1ae41a06e8a381d3a29e67ef56165