ptlrvp: fatcat: Enable PDC-driven CCD
This adds the 'pdc sbumux` command to be able to put the PDC into CCD
mode. This will need a PDC firmware update but is harmless with the
current released firmware.

With the original PDC firmware, the command will get an error:

 E: Public API command PDC_SET_SBU_MUX_MODE not sent, err=-16

With updated PDC firmware, the PDC will enter 'FORCE_DEBUG' mode as
expected.

BUG=b:406327849
TEST=tested sbumux on fatcat

Change-Id: If2bc3cfc277507757e758bf2753c30e8ceb8da85