pdc: rts54xx: fix logging of busy status
The PDC busy status is supposed to be only logged once when the PDC is
busy. In the original logic, we will set the busy bit in the cci_event.
However, in the following call of `call_cci_event_cb`, this call will
set the cci_event.raw_value to 0, as raw_value and other bits are union,
the `cci_event.raw_value` is accidentally reverted.

Move the set bit call after the `call_cci_event_cb`.

BUG=b:409727163
TEST=manually

Change-Id: Ic5c319163835e202d2437f34fdd3c7f443f3b0e9