pdc: rts54xx: fix notifying of busy status for the first time
The PDC busy status is supposed to be only notify subsystem once when
the PDC is busy. In the original logic, we will set the busy bit in the
cci_event. However, in the following call of `call_cci_event_cb`, this
call will set the cci_event.raw_value to 0, as raw_value and other bits
are union, the `cci_event.raw_value` is accidentally reverted.

Use `data->ping_retry_counter` instead to track whether it is the first
time seeing busy. Only notify the subsystem if the retry counter equals
to 1.

BUG=b:409727163
TEST=manually

Change-Id: Ic5c319163835e202d2437f34fdd3c7f443f3b0e9