uldrenite: Convert mp2964 init to use ap power hooks
The HOOK_CHIPSET_STARTUP was used to verify that the power rails were
enabled. Instead we will wait for the first AP_POWER_PRE_INIT which
also has the power rails but better aligns with the hook that's used
by the DSP comms. This ensures that this bus is used before the DSP
comms driver takes over it.

BUG=b:410646711
TEST=zmake build trulo uldrenite

Change-Id: I46c56b96acb82479316cdefda9fb790ae1ae9394