motionsense: Allow runtime disabling of motionsense
In Uldrenite we need to disable the sensor stack on the EC based on
FW_CONFIG value. If the ISH is enabled, the stack should be turned
off.

This change attempts a minimally invasive way of doing this. By
default sensors are on such that we don't change the existing
behavior. Custom board logic may call sensor_stack_runtime_disable()
in order to shut down the sensor stack (to recover, the EC must be
rebooted). This function will call the motion_sense_shutdown()
function and set an atomic variable to 0 which will terminate the
sensor loop.

Additionally, once the atomic variable is 0, none of the hooks will
work so that we don't re-configure the sensors.

BUG=b:410646823
TEST=build trulo-ish and uldrenite
TEST=./twister -s drivers.disable_motion_sense

Change-Id: I606daea83efb873ff3d6d30c3d399b7fadf586a9