switch: Do not report the WP switch over memmap
This WP_L GPIO, sourced from the GSC (or overwritten by the servo),
is dual-routed to the AP, the AP receives this write-protection
status through its GPIO subsystem. Getting the same status from
AP-EC memmap is unnecessary. So do not report it.

The following changes will remove the interrupt handlers of the WP
GPIO.

BUG=b:411207349
TEST=CQ

Change-Id: If43f0b931df9e8b489068af5d786c8864a84cb72