zephyr: npcx: infer default ec-rw memory configuration from dts
The RW size is redefined for the specific boards in the device tree. For
example, in the helipilot case, it is defined in
(ec/zephyr/program/fpmcu/helipilot/helipilot.dts). Infer the information
of CONFIG_CROS_EC_RW_SIZE directly from the device tree instead of
defining it manually in the prj.conf file, potentially causing
inconsistencies.

BUG=b:412435942
TEST=zmake build helipilot; check that CONFIG_CROS_EC_RW_SIZE has the
     correct memory size
TEST=CQ

Change-Id: I27c405cee15979e60749e2e16c6bc1196b3a6c78