zephyr: npcx: infer default wp-ro memory configuration from dts
The RO size is redefined for the specific boards in the device tree. For
example, in the helipilot case, it is defined in
(ec/zephyr/program/fpmcu/helipilot/helipilot.dts). Infer the information
of CONFIG_CROS_EC_RO_SIZE directly from the device tree instead of
defining it manually in the prj.conf file, potentially causing
inconsistencies.

BUG=b:412435942
TEST=zmake build helipilot; check that CONFIG_CROS_EC_RO_SIZE has the
     correct memory size
TEST=zmake build npcx9; check that CONFIG_CROS_EC_RO_SIZE has the
     correct memory size
TEST=CQ

Change-Id: Icb8180d2851778c06de915bda71c3775f3088a4d