zephyr: npcx: infer default EC RAM base from dts
The EC RAM base is redefined for the specific boards in the device tree. For
example, in the helipilot case, it is defined in
(ec/zephyr/program/fpmcu/helipilot/helipilot.dts). Infer the information
of CONFIG_CROS_EC_RAM_BASE directly from the device tree instead of
defining it manually in the prj.conf file, potentially causing
inconsistencies.

BUG=b:412435942
TEST=zmake build helipilot; check that CONFIG_CROS_EC_RAM_BASE has the
     correct memory address
TEST=CQ

Change-Id: I5a733d20f7675b19790b30a2de9c59a3e5bef908