zephyr: npcx: infer default available RAM size from dts
The RAM size size is redefined for the specific boards in the device
tree. For example, in the helipilot case, it is defined in
(ec/zephyr/program/fpmcu/helipilot/helipilot.dts). Infer the information
of CONFIG_CROS_EC_RAM_SIZE and CONFIG_CROS_EC_DATA_RAM_SIZE directly
from the device tree instead of defining it manually in the prj.conf
file, potentially causing inconsistencies.

BUG=b:412435942
TEST=zmake build helipilot; check that CONFIG_CROS_EC_RAM_SIZE and
     CONFIG_CROS_EC_DATA_RAM_SIZE has the correct memory size
TEST=zmake compare-builds --all
TEST=CQ

Change-Id: If7216887b2b25ebf879ab390c023195e50b3d6c1