tps6699x: Map SET_UOR ucsi command to port control register
This CL modifies the tps_6699x driver to no longer send SET_UOR as
ucsi command to the TI PDC. Instead the port control register bits
process_swap_to_dfp and process_swap_to_ufp are used instead. The
motivation for this change is that when the SET_UOR command bit
allow-dr_swap is clear, then the TI-PDC will also clear the data role
swap capable bit in the SNK/SRC cap message and this then leads to
compliance test failures.

The port control register bits are used instead to trigger data role
swaps if required, and the app config for the TI-PDC firmware is being
used to control whether data role swap requests are allowed from the
port partner which allows the flexbility to allow data role swap
requests when the port is a UFP and reject them when the port is a
DFP.

BUG=b:393455660
TEST=Validated that TD 4.11.1 complicance test is now passing and also
verified that firmware.PDDataSwap.* passes.
verified ./twister -c -T zephyr/test/pdc passes

Change-Id: Id1399f508d243a644fec9c76bf73f24607c0a26e