zephyr/test: respect size of copying response back in testing
The shim used in testing for sending host command handles copying
response data to tests. The copying process does respect the size to
copy, but copies the `args->response_max`. This causes issue if the host
command handler does not correctly specify the response size, the ztest
can still pass and failed to catch that bug.

Fix the shim by copying only the size specified in the response.

BUG=b:412803807
TEST=./twister -v -i -p native_sim -p unit_testing

Change-Id: I9323ef8eb02e9162d28733782ade24a50b3de7c6