zephyr/test: add checks for host command response size
Tests that sends host command are supposed to check the response size
as well. Add checks for most of tests for those build host command input
parameters with `BUILD_HOST_COMMAND` and `BUILD_HOST_COMMAND_RESPONSE`.

This commit skips checking on some host command calls if that call is
not using the command response, or the response size is not easy to
predict (reading blobs until the end in a loop), or need to
restructuring the test (the building of the host command args is
extracted into a utility function and is shared by various calls, and
each call have different response sizes).

BUG=b:412803807
TEST=./twister -v -i -p native_sim -p unit_testing

Change-Id: Id0c508eef88f02b295441b240db446ae1fa12990