body_detection: Make common functions ((weak))
Body detection can run on a DSP core instead of the EC. It has the
same effects but gets it's data from a different source. Allow the
function to be overridable so that ECs with DSP comms can link their
own zephyr/drivers/dsp_comms/service/remove_body_detection.cc which
provides this functionality based on the comms messages.

BUG=b:410646711
TEST=build uldrenite, flash, send fake on-body message and see that
     it's handled correctly on EC.

Change-Id: Ied83bba8a4670b2c500ea14ad1d1118a27f1162f