cec: fix wrong response data size of empty `EC_MKBP_EVENT_CEC_EVENT`
If there is no event out when `EC_MKBP_EVENT_CEC_EVENT` is received, we
should also reset the out buffer to zero, and return the size of the
empty event, so when the response is copied back to AP, there is no
garbage values returned.

Also reset the zero value to `out` variables if there is no event out
in the `EC_MKBP_EVENT_CEC_EVENT`, and return the correct size (4 bytes
for uint32_t) as data size. The comment for this test is updated
accordingly.

BUG=b:412803807
TEST=./twister -v -i -p native_sim -p unit_testing

Change-Id: Ibc553b289c49c6c279b3ce526e02069ea7102995