panic: Introduce panic_trace
Panic trace efficiently captures a trace of critical events in
uninitialized preserved ram. After a panic, the panic_trace is frozen
until it is dumped by the OS. Preserving the trace is best effort.
Integrity checks ensure corruption is detected and the trace is
discarded.

The implementation is designed to be very low overhead and reentrant.

BUG=None
TEST=Panic trace preserved across watchdog and hang. Tested on lantis
and magolor.
TEST=Pass firmware.EcStress.all

Change-Id: Ie0632bd9df5c12e47f02b0e6ce4e2dbac963cd50