motion_sense: Wake motionsense on alternate swap
When using an alternate sensor, wake the motion sense task to trigger
the ODR change so the sensors are initialized sooner.

There's a new issue that's been seen when we've added runtime disabling
of sensors. In this scenario, the EC checks the SSFC and turns off the
sensor stack. The ISH later comes online and takes over. The problem is
that the kernel queries the sensor list (using the motionsense dump
command) and at times the sensors aren't initialized yet.

The AP will likely need changes for a more robust fix where specific
error codes can be returned by the motionsense stack to notify the AP
that it's not yet ready and that additional requests should be made
later. This fix drastically reduced the number of times that the AP
reported the wrong number of sensors caused by querying too early
(before they were intialized).

BUG=b:410646158
TEST=flash uldrenite and uldrenite-ish, check that sensors work

Change-Id: I9cd7ff8f254275962062c808db5af9bedce34528