dsp_service: Use raw GPIO_TABLET_MODE_L pin value
Querying the tablet mode was returning an invalid value on uldrenite.
The original intent of the status bit TABLET_MODE was to convey the
interrupt status of the tablet mode GMR. This was always in sync with
the current tablet mode until we needed to perform a handoff of the
sensor logic from EC to ISH in uldrenite.

In this case the EC would happen to have an invalid initial state
(tablet mode), but since the ISH would start in the correct state
(clamshell) it would never update the EC.

BUG=b:410646158
TEST=flash uldrenite and uldrenite-ish, check that DUT tablet mode
     works.

Change-Id: I30f86198dbb7dc0c5b29a5beebc87db74dad781b