uldrenite: Enable force mode in suspend
Add ec-rate to all ec-s3 nodes of accelerometers. This is an override
to tell the motion sense stack that the EC needs samples at 100ms
intervals. The Kconfig CONFIG_PLATFORM_EC_SENSOR_EC_RATE_FORCE_MODE
disables interrupts when resorting to the ec-rate. In suspend, this
means that instead of getting interrupts from the lid and base at
12.5Hz each (25Hz effective wakeup) we're waking up on a timer at
10Hz.

BUG=b:410646158
TEST=flash uldrenite, run suspend_stress_test -c 10

Change-Id: I816e2a58c8c393036858085b71c89ef08c0f0f99