fatcat_it82002aw: Set GPJ4 high on init
The smc_extsmi_n_c_ec_pch_int signal is an indication to the AP that
there are EC events to be processed. This needs to be deasserted on boot
to avoid false requests. Note that this pin is marked ACTIVE_HIGH even
though it is active low because it will be managed by classic cros-ec
code which does its own inversion. The GPIO_EC_INT_L label will be
needed in a follow-up patch when we enable MKBP_USE_GPIO_AND_HOST_EVENT.

Updated GPIO spreadsheet and regenerated with:

  arb export-ec-devicetree-csv
    --sheet-url=$sheet
    --subsheet='EC Pinout'
    --first-row=3
    $schematics

BUG=b:414316209
TEST="HC 0x0067 err 9" messages can be triggered by toggling
	smc_extsmi_n_c_ec_pch_int_odl_r. no HC spam on boot.

Change-Id: I899844fd37bbf5229d284dc1c8cded1bd6531972