tps6699x: Set get_pd_message in GET_CAPABILITY response
Because the TI-PDC supports UCSI_GET_PD_MESSAGE, the bit
get_pd_message should be set in bmOptionalFeatures field of the
GET_CAPABILITY response. The ucsi kernel driver will not send
UCSI_GET_PD_MESSAGE, which is required for getting discovery results,
if this bit is not set. This CL adds a change to intercept the
response and set the bit in the driver as a workaround until this
issue is fixed in TI-PDC FW.

BUG=b:414863461
TEST=Verified that the the kernel is now sending UCSI_GET_PD_MESSAGE
and discovery values are populated in the kernel.

Change-Id: I7ad99c8658af93c80859e60f11a0d70bd43f2548