tps6699x: Save command for set_sink_path/set_rdo callback
The tps6699x adds logic to both the SET_SINK_PATH and SET_RDO commands
to ensure the commands won't be rejected by the PDC. Part of this
logic includes a callback function to trigger sending the command if
the conditions weren't met when initially send by the PDM state
machine.

The callback function was relying on the value data->cmd to be
preserved, but this value is cleared in st_idle_entry. This CL adds a
variable delayable_cmd to store the command which initiated the
callback so the callback will sned the desired command. Not sending
SET_SINK_PATH from the callback results in ~2.5 extra dealy before the
kernel is notified that battery charging is active.

BUG=b:404286535
TEST=Verified that SET_SINK_PATH (task_srdy) is called after it's been
requested by the PDM state machine and the PDC reports that a valid
USB-PD contract is in place. Also observed that the battery icon on
the UI indicates charging now within ~2 seconds of a charger being
connected. Previously, this would take upwards of 5 seconds.

Change-Id: I1c0787479c9f5773cc815095125f6f74b0d7b725