zephyr: pass CROSTC_ADD_IMPLICIT_CFLAGS_FOR env var though
In b/403315166, we adopted a flow to allow users to fix 'fatal' warnings
in builds that get introduced by toolchain rolls. The tl;dr of this flow
is that:

1. New fatal warnings are automatically suppressed during new toolchain
   version testing.
2. The suppressed warnings turn into a bundle of implicit `-Wno-error`
   flags, based on `CROSTC_ADD_IMPLICIT_CFLAGS_FOR` (env) or
   `-_DCROSTC_ADD_IMPLICIT_CFLAGS_FOR` (cflag).
3. The new toolchain lands and is verified to stick.
4. Owners of packages with these warnings get bugs notifying them of the
   suppressions (with simple instructions for how to fix the
   suppressions).

Since zmake filters env and embedded projects try to avoid target
CFLAGS, add the requisite env var to the allowlist.

BUG=b:406012709
TEST=emerge

Cq-Depend: chromium:6546704
Change-Id: Ib8d260416f5d929e55abcc89dbbf2f5fd186cf28