test/run_device_tests: Update expected power for dragonclaw v4 boards
The numbers for MCU are higher and the numbers for the sensor are lower.
I adjusted these values based on ToT measurments, but it doesn't appear
to be a regression. I also get higher numbers for the MCU measurement
when using the same version of the firmware that we measured with
dragonclaw v3:

BUG=b:358679133
TEST=Dragonclaw v4:
     ./test/run_device_tests.py --board bloonchipper \
       -t power_utilization_sleep
TEST=Dragonclaw v4:
     ./test/run_device_tests.py --board bloonchipper \
       -t power_utilization_idle
TEST=Dragonclaw v4:
     ./test/run_device_tests.py --board bloonchipper \
       --zephyr -t power_utilization_sleep
TEST=Dragonclaw v4:
     ./test/run_device_tests.py --board bloonchipper \
       --zephyr -t power_utilization_idle

Change-Id: Iadfdf60cbac51708ced04a4976c76e63ccd77b1b