bluey: Add long warm reset sequence for QC_EXP SoC
This patch adds long warm reset sequence for QC_EXP SoC.

Upon SYS_RST_L assertion, a 20ms timer begins to differentiate between a
short and long warm reset.
If SYS_RST_L deasserts before the timer expires, the timer is cancelled,
and an EC-initiated short warm reset sequence is performed. However, if
the 20ms timer expires (meaning SYS_RST_L remained asserted for at least
20ms), a long warm reset sequence is initiated. This sequence involves
powering off the system and waiting for SYS_RST_L deassertion. Upon
deassertion, the system performs a power-on sequence if it was
previously in S0 state.

This patch adds a interrupt for the SYS_RST_ODL signal and updates the
signal as an input to the EC.

BUG=b:408278047, b:415184416
TEST=zmake build bluey

Change-Id: I72d4112617e7dfd320a6dd14c1ff71118af82ffd