pdc: Support flashing RTK PDCs in ROM mode using console PDC updater
Update the EC console-based RTK PDC updater to support flashing the PDC
if it is in ROM mode. This is detected based on the GET_IC_STATUS
response. When the PDC is in ROM mode, perform an erase of flash before
flashing the new image, and skip the validation step as it is not
supported.

BUG=b:417489560
TEST=Erase a PDC, and reflash it.

Cq-Depend: chromium:6544873
Change-Id: I5a75fb6b257036b71eb18359eea708264abff7b1