ap_pwrseq: Make S0ix entry timeout adjustable
The S0ix entry timeout is currently set by the legacy config item
CONFIG_SLEEP_TIMEOUT_MS, which does not have a Kconfig option to
override in the Zephyr EC. Intel needs to be able to adjust this value.

Remove this config reference from the driver and instead use a property
from the devicetree (s0ix-sleep-timeout under an intel,ap-pwrseq node).
The DT property has the same default value (10s)

Set this DT prop to 20 sec in the PTL-RVP.

BUG=b:417404965
TEST=zmake build fatcat_npcx9m7f
TEST=./twister -T zephyr/test/ap_power

Change-Id: I5838936d52a2857e70219f0960e6d4a3a3480221