body_detect: Notify host only in S0
When the DUT is in a suspended state S0ix we don't need to notify the
host of the change as the thermals are not a concern in suspend. This
allows the DUT to remain suspended longer and saves power.

To achieve this we need to expose some motion_sense values/functions
that were prioviously only exposed for tests.

BUG=b:410646158
TEST=flash uldrenite, run suspend_stress_test -c 10
TEST=flash uldrenite, powerd_dbus_suspend, rotate screen past 180,
     check that it woke up the DUT

Change-Id: Iab654795791402422e2832a6491efe0da43af7f9