uldrenite: Fix current devices without MP2964 rework
When testing the current devices with ISH enabled the MP2964 was moved
to i2c5_0. When the write fails since the pins aren't connected it's
possible that the bus errors (I believe due to a floating pin). Move
back to the i2c1_0, revert the rail_a changes, and set the init back
to AP_POWER_PRE_INIT. This effectively reverts chromium:6486335.

BUG=b:415650413
TEST=Check uldrenite passes: tablet mode, screen rotation, suspend
     stress tests, and ec reboot

Change-Id: I4c4803193460c7535fb1aa6bef0e32ea8503b094