twister_launcher: use a static list to find modules
Discovering modules automatically from twister runs can cause issues
when modules are added to the repository but the files needed to support
them have not being downstreamed yet. This caused ci breakages in the
past.

Drop the module detection code and use a static list instead, the normal
EC build uses a static list of modules anyway so having a static list
here does not really add much in term of tech debt.

BUG=b:418319903
TEST=./twister

Change-Id: I752269c070ccaea4431a7a81ce86a2b6e7bff14b