dsp_comms: Fix Nuvoton i2c error
The Nuvoton i2c controller has an issue where if we do a warm reset
while we're in target mode, the initialization logic fails and puts
the bus in a bad state. The only way we've found to recover is to
first recover the bus, then write a byte to it. Note that just
recovery and just a write aren't sufficient.

BUG=b:416656289
TEST=On uldrenite call ectool reboot_ec
TEST=On uldrenite check screen rotation works
TEST=On uldrenite check tablet mode works
TEST=On uldrenite run "crash unaligned" in EC console

Change-Id: I67266382452b37a4fd25246149d54413023dce6e