CBI: Deprecate FACTORY_CALIBRATION_DATA and COMMON_CONTROL
The FACTORY_CALIBRATION_DATA tag was used in a canceled project to store
the calibration data for a PWM brightness. No project uses it now.

The COMMON_CONTROL tag was used for runtime configuration to enable
features, like BCIC or USCI. However, runtime enabling of BCIC or UCSI
was no longer a requirement.

So these 2 tags can be deprecated.

BUG=b:415806092
TEST=make runhosttests -j

Change-Id: I7fe57c2ae830d97162a154a5f7ba50fa9a7fe0c3