pdc: tps6699x: Add support for retimer enable/disable
This CL adds a new PDC API command CMD_PDC_SET_AP_POWER_STATE which is
used to inform the PDC of the current AP power state to support
enabling/disabling 3rd party retimer chips. In addition this CL adds
support for this command to the tps6699x driver.

BUG=b:378956877
TEST=Validated that  CMD_PDC_SET_AP_POWER_STATE is sent following AP
power state changes and that register 0x20 in the TI-PDC is updated as
expected.
LOW_COVERAGE_REASON=b/378956877 Feature is needed for bringup. Will
add test coverage when I am back

Change-Id: I7709a53dafb03112f67915f5d8eb5703c389c5e3