battery: report communication errors through ACPI
Currently the EC reports errors while retrieving dynamic battery
information to the host by setting EC_BATT_FLAG_INVALID_DATA (shared
with ACPI firmware by setting the BFIV bit in shared memory), but ACPI
firmware never reads that bit so if there is an error reading any
dynamic data from the battery the fields that fail will always read as
zero and the OS receives no indication of error.

The frequency of battery communication errors is surprisingly high, with
around 20000 apparent individual errors found in user feedback from a
90-day period of time. This probably represents a lower bound for the
actual frequency of errors, because (as explained below) they are
typically invisible to the user and will only be captured incidentally
when users opt to share feedback about something else. For some
platforms, the battery fuel gauge vendors have confirmed that it's
expected the battery will sometimes NAK I2C transactions when it's busy
doing internal housekeeping so it seems clear that battery errors are
not uncommon.

This change makes host battery reporting set all dynamic fields to a new
"invalid" value if there is an error while reading from the battery,
avoiding incorrect 0 values that can easily be misinterpreted. For ACPI
platforms, this will cause the kernel to return ENODEV to userspace when
attempting to read the affected properties so the OS can choose to
handle these issues in whatever way is desired. Although parts of the
charge manager have finer-grained flags for which fields contain invalid
data, those are not synchronized with the battery state reporting code
so it must take a pessimistic interpretation that assumes all dynamic
fields contain invalid data.

The behavior of reading all-zero on error has historically not been an
issue because ChromeOS powerd includes a number of checks for whether
battery information retrieved from sysfs appears valid, and it will
simply ignore battery state snapshots that seem wrong. In particular, it
assumes that any total charge level of 0 is false so a transient battery
communication error (where the battery still appears to be present but
is unreponsive) will be invisible to consumers of battery state provided
by powerd. Since errors of this kind are papered over in powerd, other
OSes (that do not use powerd) will generally see incorrect state of
charge and can react badly; possibly by shutting down even though the
battery is not critically low, or incorrectly notifying the user of low
battery.

This change primarily affects ACPI behavior, because non-ACPI platforms
do not communicate with the EC via shared memory map: they generally use
I2C from the host processor to EC via the virtual battery feature, and
the virtual battery implementation already inspects battery flags to
return errors for fields that were not successfully read from the
battery.  For platforms where the host communicates directly with the
battery (usually by tunneling through the EC), communication errors
should directly bubble up to the host OS so no changes are required.

BUG=b:410996664
TEST=When battery communication is forced to fail but the battery is
     still reported as present, reads from
     /sys/class/power_supply/BAT0/capacity on redrix fail with ENODEV.
TEST=On ChromeOS, powerd still ignores battery readings taken when the
     battery is returning errors because it assumes they're wrong so
     there is no user-visible effect from this change.

Change-Id: If7eadd626bf409d885e3427664b4cf7a759d06a6