panic_data: Expand panic_data flags to 16 bits
The panic_data flags field is only bits and all 8 bits are claimed. The
reserved field following the flags field is unused. Merge flags field
and reserved field to make flags 16 bits.

Increment panic_data struct version from 2 to 3.

When interpreted as little-endian the v2 is backwards compatiable with v2.

BUG=b:420744645
TEST=Build

Change-Id: I7612329af818d09ea94522d95cd67c3db96ee580