Change cbi system locked behavior
Instead of blocking writes to board and oem_id, block init always, and
writes to those 2 fields if they are already present. If board or oem_id
are missing, then go ahead and allow them to be set.

For devices with cbi in flash (like brox), if the CBI gets cleared
accidently while flashing ec, there is no way to fix it.

Add test case with system locked.

Remove init flag from tests that seem to believe it means "Force a
reload" instead of "clear all the cbi data".

BUG=None
TEST=Adjusted unit tests only

Change-Id: I94d2c2b421ed6c1aaafc7880c3286c0bd1399e15