zephyr/panic: Infer CFSR from error reason
The CFSR register is not part of the esf and is cleared before
k_sys_fatal_error_handler is called. The CFSR is translated to
k_fatal_error_reason_arch and passed as reason.

There is no field in cortex-m panic_data to store a generic reason. For
software panics registers is overloaded to store the reason.

This change makes a best effort to infer CFSR from the passed in reason.
By default CFSR and all other special registers are set to -1 to make it
clear these registetrs are not valid.

BUG=b:245423691
TEST=CFSR correct after `crash unaligned`

Change-Id: Ie64755b7ef34f1d5e172c8c62f566b630b979f1f