pdc: tps6699x: Move register init to ST_INIT
Initialization of several key registers (interrupts, port control, etc)
is currently done in pdc_init which is only called once when the driver
is initialized. If the PDC resets for any reason without the EC
resetting, this will result in the driver getting into a state where it
is not configured correctly.

This commit makes the following changes:
* Moves critical register initialization to `st_init_run`
* Delays `init_done` until after UCSI notifications have been configured
* Adds a retry mechanism to initialization
* GAID task pushes the system back into INIT state after 1.6s delay

BUG=b:421342154, b:422062261
TEST=./twister -cv --coverage -T zephyr/test/pdc/

Change-Id: Iae1912185a4ec9e085d6023a7e413326e4036548