gsctool: Print request reboot error to stdout
When using gsctool via devboard service, the stderr is not easily
accessible in tast tests. For NT images, the update failure for board
id mismatch doesn't occur until the reset request, and we want to be
able to detect that in tast when running with new `--skip_bid_check`
to validate that RW FW is preventing bad bid updates.

BUG=none
TEST=verify that tast can detect bid update failure from Ti50 FW
TEST=BED=DT make tast TAST_EXPR=gscdevboard.GSCFactoryUpdate.ti50_0_21_1

Change-Id: Ia4391ac41bb174e306bea826ed1391d26b0eaca8