zephyr: helipilot: Correct flash configurations for Helipilot baseboard
The commit updates the `prj.conf` for the helipilot baseboard to ensure
its flash operations are correctly configured. First, we are enabling
`CONFIG_FLASH_EX_OP_ENABLED`. This setting is crucial and is required
for certain flash operations to function properly on this specific
platform. Second, we are disabling
`CONFIG_PLATFORM_EC_USE_ZEPHYR_PAGE_LAYOUT`. This configuration, which
is closely tied to `CONFIG_FLASH_MULTIPLE_REGION`, is specifically
designed to accomodate the unique flash organization of STM32 chips.
Since helipilot baseboard does not utilize an STM32 chip but rather an
NPCX9 one, this layout is unncessary and can lead to incorrect
bahaviour. Disabling it ensures that helipilot baseboard uses the
appropriate, general-purpose flash configuration.

BUG=b:419592396, b:421211718, b:422120872
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t
     system_is_locked.* --zephyr
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t
     flash_write_protect --zephyr

Change-Id: Ied6c1ea40e7f9115a7a3fb0f600bd1940f8957a8