keyboard: Add a stub keyboard protocol
Add a new type of keyboard protocol that can be selected in place of
MKBP or 8042 that discards all the key events instead of trying to send
them to the AP. This is helpful for testing and bringup when an AP might
not be enabled yet and reduces the amount of dependencies required to
perform simple keyboard validation.

This change removes the general CONFIG_AP requirement for keyboard
protocols (MKBP and 8042 still require it) and also permits the Vivaldi
keyboard driver to be enabled with 8042 *or* the stub driver.

BUG=b:422781395
TEST=zmake build ocelotrvp*
TEST=Run on WCL-RVP
LOW_COVERAGE_REASON=Write test for keyboard code in b/422821421

Change-Id: I9576db12c60969a036a5044b05fc9db49d405697