keyboard: Add a stub keyboard protocol
Add a new type of keyboard protocol that can be selected in place of
MKBP or 8042 that discards all the key events instead of trying to send
them to the AP. This is helpful for testing and bringup when an AP might
not be enabled yet and reduces the amount of dependencies required to
perform simple keyboard validation.

This change removes the general CONFIG_AP requirement for keyboard
protocols (MKBP and 8042 still require it) and also permits the Vivaldi
keyboard driver to be enabled with 8042 *or* the stub driver.

The keyboard protocol is also made optional since some devices like
roach don't use any of the three choices but provide their own from
board code. Because 8042 is no longer the default in this case,
select 8042 in a few boards that did not explicitly specify this (e.g.
Skyrim and Trulo designs) but are using 8042. Make similar updates to
tests.

BUG=b:422781395
TEST=zmake compare-builds -a
TEST=zmake build ocelotrvp*
TEST=Run on WCL-RVP
LOW_COVERAGE_REASON=Write test for keyboard code in b/422821421

Change-Id: I9576db12c60969a036a5044b05fc9db49d405697
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6625971
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
(cherry picked from commit 049dfa36b1058a6e57c4671ab4b74d9714239d90)