intelrvp: Fix charge seeding from barrel jack
The barrel jack code was unconditionally seeding all charger ports and
suppliers with 0A/5V.  This caused the "chgsup" command to display
confusing information.

The charge manager now correctly seeds charger suppliers when a
dedciated charge port is present, so the seeding done by the intelrvp
is not needed.

This also fixes the charge manager status when the barrel jack is
disconnected by setting the supplier status to 0A/0V.

BUG=none
TEST=Boot ocelotrvp-npcx - run "chgsup" command

Change-Id: Ia391b10a8ef1ff76a363a1153ea438b4d892f9bd