gsctool: Convert from BE transport to host encoding
The TPMV standard is to transport words at BE. Fix issue when leaving
the board_id struct fields in their transport form instead of converting
them to the host endianness.

My initial testing with tast was invalid because I performed the last
round of testing with the `--skip_bid_check` flag mistakenly.

BUG=b:416600474,b:423619250
TEST=See that test image with `TEST` RLZ is appropriately constrained
     with a DT and NT GSC. Both through a manual test and through
     tast without the pending tast CL that add `--skip_bid_check`
TEST=BED=DT make tast TAST_EXPR=gscdevboard.GSCFactoryUpdate.ti50_0_21_1

Change-Id: I9013e5a48a2d29097bdb00cd1b9c1b9567ff2a02