zmake: Allow specifying projects by program directory name
Add a new feature when selecting projects for `zmake build | configure |
list-projects | compare-builds` to simply specifying multiple projects
to build. When expressions starting with `%` are passed in, select all
projects that are declared in a directory matching that expression.

For example:

    # Build all nissa variants
    zmake build %nissa

    # Build all skyrim and skywalker boards
    zmake build %sky*

Specific project names, wildcard project names, and %program names can
all be combined and used multiple times. The final set of projects to be
built is the union of all expressions.

Update tests to make sure this feature works correctly.

Add an assertion so that actual project names do not start with '%'

BUG=None
TEST=zmake build ...
TEST=zmake/run_tests.sh

Change-Id: Icabe45b955963ac4a5bfa7f5bef4de438da8bacc