ptlrvp_mchp: Change SPI NOR init priority for MCHP RVP
Microchip RVP uses the SPI NOR which was default set to init very
late in the boot process. For RVP this needs to be init after the SPI
is init (priority 50) and before the cros shim is init (priotity 55).
With that in mind the priority is set to 54 which is the latest the SPI
NOR can be init while still allowing the system to boot.

BUG=b:423808174
TEST=zmake build ptlrvp_mchp, verify system boots to kernel

Change-Id: I4d394b39ded8020784fe32521a8265dfcfca1d72